+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Tamura 2023 remdesivir / GS-441524 ([doi:10.1002/psp4.12936](https://doi.org/10.1002/psp4.12936)) -- Japanese adults hospitalised with moderate-to-severe COVID-19.
+- Add Frederiksen 2023 brexpiprazole with its DM-3411 and DM-3412 metabolites ([doi:10.1002/cpt.2791](https://doi.org/10.1002/cpt.2791)) -- healthy subjects and patients with schizophrenia, major depressive disorder, or attention deficit hyperactivity disorder.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.